analysis/requirements.txt
supabase>=2.4.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1
Pillow>=10.3.0
typer[all]>=0.12.3
pytesseract>=0.3.10
rich>=13.7.1
tqdm>=4.66.2
